+++ src/main/java/com/udacity/jwdnd/course1/cloudstorage/controller/NoteController.java
@@ -1,10 +1,12 @@
 package com.udacity.jwdnd.course1.cloudstorage.controller;
 
 import com.udacity.jwdnd.course1.cloudstorage.DTO.NoteDTO;
+import com.udacity.jwdnd.course1.cloudstorage.model.Note;
 import com.udacity.jwdnd.course1.cloudstorage.model.ResultMessage;
 import com.udacity.jwdnd.course1.cloudstorage.model.enums.ResultMessageType;
 import com.udacity.jwdnd.course1.cloudstorage.services.NoteService;
 import com.udacity.jwdnd.course1.cloudstorage.services.UserService;
+import com.udacity.jwdnd.course1.cloudstorage.services.UtilService;
 import org.springframework.security.core.Authentication;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
@@ -16,10 +18,12 @@ public class NoteController {
 
     private final NoteService noteService;
     private final UserService userService;
+    private final UtilService utilService;
 
-    public NoteController(NoteService noteService, UserService userService) {
+    public NoteController(NoteService noteService, UserService userService, UtilService utilService) {
         this.noteService = noteService;
         this.userService = userService;
+        this.utilService = utilService;
     }
 
     @PostMapping("/add")
@@ -35,13 +39,17 @@ public class NoteController {
                 message.setMessage("Note added successfully.");
             }
         } else {
-            int result = noteService.updateNote(noteDTO);
-            if (result < 1) {
-                message.setMessageType(ResultMessageType.ERROR);
-                message.setMessage("There was an error updating the note.");
-            } else {
-                message.setMessageType(ResultMessageType.SUCCESS);
-                message.setMessage("Note updated successfully.");
+            Note note = noteService.getNotById(noteDTO.getNoteId());
+
+            if (utilService.checkSameUser(note.getUserId(), auth)) {
+                int result = noteService.updateNote(noteDTO);
+                if (result < 1) {
+                    message.setMessageType(ResultMessageType.ERROR);
+                    message.setMessage("There was an error updating the note.");
+                } else {
+                    message.setMessageType(ResultMessageType.SUCCESS);
+                    message.setMessage("Note updated successfully.");
+                }
             }
         }
         model.addAttribute("resultMessage", message);
@@ -49,16 +57,26 @@ public class NoteController {
     }
 
     @GetMapping("/delete/{noteId}")
-    public String deleteNote(@PathVariable("noteId") int noteId, Model model) {
+    public String deleteNote(@PathVariable("noteId") int noteId, Model model, Authentication auth) {
         ResultMessage message = new ResultMessage();
-        int deleted = noteService.deleteNote(noteId);
-        if (deleted < 1) {
-            message.setMessageType(ResultMessageType.ERROR);
-            message.setMessage("There was an error deleting de note.");
+
+        Note note = noteService.getNotById(noteId);
+
+        if (utilService.checkSameUser(note.getUserId(), auth)) {
+            int deleted = noteService.deleteNote(noteId);
+            if (deleted < 1) {
+                message.setMessageType(ResultMessageType.ERROR);
+                message.setMessage("There was an error deleting de note.");
+            } else {
+                message.setMessageType(ResultMessageType.SUCCESS);
+                message.setMessage("The note was deleted successfully.");
+            }
         } else {
-            message.setMessageType(ResultMessageType.SUCCESS);
-            message.setMessage("The note was deleted successfully.");
+            message.setMessageType(ResultMessageType.ERROR);
+            message.setMessage("You can't delete another user's credential.");
         }
+
+
         model.addAttribute("resultMessage", message);
         return "/result";
     }
